Fix for TRANSFER transactions
This changeset proposes a change that fixes and closes #2. There was
an error made writing the code that serializes a transaction. As a
result, it always sets the operation to `CREATE` instead of allowing
it to be set to `TRANSFER`.

@NikitaLos originally filed the bug but we fixed it before noticing
his issue.

Additionally, a fix for #3 has been applied, by applying an `@Ignore`
to the failing test, so the build will pass for now.

In my case I was missing Jadler (so I filed #11), and this consequently
fixes and closes #11 as well.

To sum up:
- [x] Fixes and closes #11 (Jadler test dependency missing)
- [x] Fixes and closes #3 (mvn clean install failing due to failing tests)
- [x] Fixes and closes #2 (Transaction operation issue)